Dynamic load media section to avoid SSR hydration error due to client Posthog flag
