+++ frontends/main/next.config.js
@@ -84,13 +84,12 @@ const nextConfig = {
           },
         ],
       },
-
       {
         source: "/:path*",
         headers: [
           {
             key: "Permissions-Policy",
-            value: "picture-in-picture=all",
+            value: "picture-in-picture=*",
           },
         ],
       },
